ve_image_sync.py attempts to upload an image as public by default
Issues:
Fixes #24

Problem:
If Heat template attempts to upload an image to Glance and that user
doesn't have the permissions to upload an image as public, the upload
will fail. We need to parameterize this functionality and allow a user
to set it as public if they wish. The default behavior will be as
non-public. This will be just a boolean.

Analysis:
Created a new argparse argument that is stored as False by default and
True is the user includes it, '--public-image' or '-p'.

Tests:
Unit tests pass